add: user logic

+++ src/main/java/rugbyniela/entity/dto/userSeasonScore/UserCoalitionHistoryResponseDTO.java
@@ -3,7 +3,7 @@ package rugbyniela.entity.dto.userSeasonScore;
 public record UserCoalitionHistoryResponseDTO(
 		String coalitionName,
 	    String seasonName,  
-	    Double score,
+	    int score,
 	    Long seasonId) {
 
 }
